allocate Prophet in separate file
