+++ tests/test_models.py
@@ -11,7 +11,8 @@ from tot.datasets.dataset import Dataset
 from tot.evaluation.metrics import ERROR_FUNCTIONS
 from tot.models.models_naive import NaiveModel, SeasonalNaiveModel
 from tot.models.models_neuralprophet import NeuralProphetModel, TorchProphetModel
-from tot.models.models_simple import LinearRegressionModel, ProphetModel
+from tot.models.models_prophet import ProphetModel
+from tot.models.models_simple import LinearRegressionModel
 
 log = logging.getLogger("tot.test")
 log.setLevel("WARNING")
